LPS-120808 Move isMapped to utils
