+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/fragment-content/FragmentContentInteractionsFilter.js
@@ -23,6 +23,7 @@ import selectCanUpdateEditables from '../../selectors/selectCanUpdateEditables';
 import selectCanUpdatePageStructure from '../../selectors/selectCanUpdatePageStructure';
 import {useSelector, useSelectorCallback} from '../../store/index';
 import {deepEqual} from '../../utils/checkDeepEqual';
+import isMapped from '../../utils/isMapped';
 import {useToControlsId} from '../CollectionItemContext';
 import {
 	useActivationOrigin,
@@ -36,7 +37,6 @@ import {
 } from '../Controls';
 import {useSetEditableProcessorUniqueId} from './EditableProcessorContext';
 import {getEditableElement} from './getEditableElement';
-import isMapped from './isMapped';
 
 const EDITABLE_CLASS_NAMES = {
 	active: 'page-editor__editable--active',
