+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/MappingSelector.js
@@ -19,13 +19,13 @@ import React, {useEffect, useState} from 'react';
 
 import {addMappedInfoItem} from '../../app/actions/index';
 import {useCollectionConfig} from '../../app/components/CollectionItemContext';
-import isMapped from '../../app/components/fragment-content/isMapped';
 import {EDITABLE_TYPES} from '../../app/config/constants/editableTypes';
 import {LAYOUT_TYPES} from '../../app/config/constants/layoutTypes';
 import {config} from '../../app/config/index';
 import CollectionService from '../../app/services/CollectionService';
 import InfoItemService from '../../app/services/InfoItemService';
 import {useDispatch, useSelector} from '../../app/store/index';
+import isMapped from '../../app/utils/isMapped';
 import {useId} from '../../app/utils/useId';
 import ItemSelector from './ItemSelector';
